Revert "Issue 536: Refactored urls in tests"
This reverts commit b2842525798ba8f8e6a04624f99fb33a07796353.
